+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 1.1.7
 
 - Added example file `example_streamed_request.dart`.
-- Updated dependency on `multitasking` package to version 5.4.0.
+- Updated dependency on `multitasking` package to version 5.6.0.
 
 ## 1.1.6